Split state changes into multiple files
