+++ lib/commands/project_state_changes.dart
@@ -1,0 +1,72 @@
+/*
+  Copyright (C) 2022 Joshua Wade
+
+  This file is part of Anthem.
+
+  Anthem is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  Anthem is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+  General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with Anthem. If not, see <https://www.gnu.org/licenses/>.
+*/
+
+part of 'state_changes.dart';
+
+abstract class ProjectStateChange extends StateChange {
+  int projectID;
+
+  ProjectStateChange({required this.projectID});
+}
+
+/*
+ * State changes
+ */
+
+class GeneratorAdded extends GeneratorStateChange {
+  GeneratorAdded({
+    required int projectID,
+    required int generatorID,
+  }) : super(projectID: projectID, generatorID: generatorID);
+}
+
+class GeneratorRemoved extends GeneratorStateChange {
+  GeneratorRemoved({
+    required int projectID,
+    required int generatorID,
+  }) : super(projectID: projectID, generatorID: generatorID);
+}
+
+class ActiveGeneratorSet extends GeneratorStateChange {
+  ActiveGeneratorSet({
+    required int projectID,
+    required int? generatorID,
+  }) : super(projectID: projectID, generatorID: generatorID);
+}
+
+class PatternAdded extends PatternStateChange {
+  PatternAdded({
+    required int projectID,
+    required int patternID,
+  }) : super(projectID: projectID, patternID: patternID);
+}
+
+class PatternDeleted extends PatternStateChange {
+  PatternDeleted({
+    required int projectID,
+    required int patternID,
+  }) : super(projectID: projectID, patternID: patternID);
+}
+
+class ActivePatternSet extends PatternStateChange {
+  ActivePatternSet({
+    required int projectID,
+    required int? patternID,
+  }) : super(projectID: projectID, patternID: patternID);
+}
